+++ karma/Components/Dialogs/DialogAddNewAnnouncement.razor
@@ -11,53 +11,47 @@
     </TitleContent>
 
     <DialogContent>
-        <MudTextField
-            @bind-Value="@_title"
-            Label="Title"
-            Required="true"
-            RequiredError="Title is required" />
-        
-        <MudTextField
-            @bind-Value="@_description"
-            T="string"
-            Label="Description"
-            Variant="Variant.Text"
-            Lines="15"
-            Required="true"
-            RequiredError="Description is required"/>
+        <MudTextField @bind-Value="@_title"
+                      Label="Title"
+                      Required="true"
+                      RequiredError="Title is required" />
+
+        <MudTextField @bind-Value="@_description"
+                      T="string"
+                      Label="Description"
+                      Variant="Variant.Text"
+                      Lines="15"
+                      Required="true"
+                      RequiredError="Description is required" />
     </DialogContent>
 
     <DialogActions>
-        <MudButton
-            Variant="Variant.Outlined"
-            StartIcon="@Icons.Material.Filled.Save"
-            Color="Color.Success"
-            Size="Size.Medium"
-            @onclick="AddAnnouncement">
-        Save
+        <MudButton Variant="Variant.Outlined"
+                   StartIcon="@Icons.Material.Filled.Save"
+                   Color="Color.Success"
+                   Size="Size.Medium"
+                   @onclick="AddAnnouncement">
+            Save
         </MudButton>
 
-        <MudButton
-            Variant="Variant.Outlined"
-            StartIcon="@Icons.Material.Filled.Cancel"
-            Color="Color.Error"
-            Size="Size.Medium"
-            @onclick="Cancel">
-        Cancel
+        <MudButton Variant="Variant.Outlined"
+                   StartIcon="@Icons.Material.Filled.Cancel"
+                   Color="Color.Error"
+                   Size="Size.Medium"
+                   @onclick="Cancel">
+            Cancel
         </MudButton>
 
-        <InputFile
-            id="fileInput"
-            OnChange="UploadImage"
-            hidden multiple accept=".jpg, .jpeg, .png" />
+        <InputFile id="fileInput"
+                   OnChange="UploadImage"
+                   hidden multiple accept=".jpg, .jpeg, .png" />
 
-       <MudButton
-            HtmlTag="label"
-            Variant="Variant.Filled"
-            Color="Color.Primary"
-            StartIcon="@Icons.Filled.CloudUpload"
-            for="fileInput">
-        Upload your images
+        <MudButton HtmlTag="label"
+                   Variant="Variant.Filled"
+                   Color="Color.Primary"
+                   StartIcon="@Icons.Filled.CloudUpload"
+                   for="fileInput">
+            Upload your images
         </MudButton>
     </DialogActions>
 </MudDialog>
@@ -65,7 +59,7 @@
 @code
 {
     [CascadingParameter] MudDialogInstance MudDialog { get; set; }
-    
+
     private string _title { get; set; }
     private string _description { get; set; }
     private byte[] _byteArray { get; set; }
@@ -81,9 +75,8 @@
             !String.IsNullOrWhiteSpace(_description) &&
             _byteArray != null)
         {
-            Announcement announcement = new Announcement{ Title = _title, Description = _description, Img = _byteArray};
+            Announcement announcement = new Announcement { Title = _title, Description = _description, Img = _byteArray };
 
-            Snackbar.Add("Announcement added!", Severity.Success);
             MudDialog.Close(DialogResult.Ok(announcement));
         }
     }
