+++ karma/Components/Dialogs/DialogAddNewListing.razor
@@ -61,7 +61,6 @@
         {
             Listing listing = new Listing{ Title = _title, Description = _description };
 
-            Snackbar.Add("Listing added!", Severity.Success);
             MudDialog.Close(DialogResult.Ok(listing));
         }
     }
