+++ karma/Pages/Charities.razor
@@ -1,14 +1,13 @@
 ﻿@page "/charities"
 @inject IDialogService _dialogService
-
+@inject ISnackbar Snackbar
 <link rel="stylesheet" href="@(nameof(karma)).styles.css" />
 
 <div class="fab">
-    <MudFab
-        Color="Color.Primary"
-        Icon="@Icons.Material.Filled.Add"
-        Label="Add charity"
-        @onclick="OpenDialog" />
+    <MudFab Color="Color.Primary"
+            Icon="@Icons.Material.Filled.Add"
+            Label="Add charity"
+            @onclick="OpenDialog" />
 </div>
 
 <div class="charities-place">
@@ -24,11 +23,11 @@
                 <!--REQUIREMENT 1.9-->
                 @foreach (var charity in _charities)
                 {
-                <li>
-                    @if (MainUser.IsAdmin == true)
-                    {
-                        <img class="delete-btn" src="img/delete-btn.png" @onclick="() => OpenDialogConfirmation(charity.Id)">
-                    }
+                    <li>
+                        @if (MainUser.IsAdmin == true)
+                        {
+                            <img class="delete-btn" src="img/delete-btn.png" @onclick="() => OpenDialogConfirmation(charity.Id)">
+                        }
                         <img src="img/save-the-children.png" alt="Charity logo">
                         <h2>@charity.Name</h2>
                         <p>@charity.Description</p>
