+++ karma/Pages/Listings.razor
@@ -1,14 +1,13 @@
 ﻿@page "/listings"
 @inject IDialogService _dialogService
-
+@inject ISnackbar Snackbar
 <link rel="stylesheet" href="@(nameof(karma)).styles.css" />
 
 <div class="fab">
-    <MudFab
-        Color="Color.Primary"
-        Icon="@Icons.Material.Filled.Add"
-        Label="Add listing"
-        @onclick="OpenDialog" />
+    <MudFab Color="Color.Primary"
+            Icon="@Icons.Material.Filled.Add"
+            Label="Add listing"
+            @onclick="OpenDialog" />
 </div>
 
 <div class="listings-place">
@@ -22,11 +21,11 @@
             <ul>
                 @foreach (var listing in _listings)
                 {
-                <li>
-                    @if (MainUser.IsAdmin == true)
-                    {
-                        <img class="delete-btn" src="img/delete-btn.png" @onclick="() => OpenDialogConfirmation(listing.Id)">
-                    }
+                    <li>
+                        @if (MainUser.IsAdmin == true)
+                        {
+                            <img class="delete-btn" src="img/delete-btn.png" @onclick="() => OpenDialogConfirmation(listing.Id)">
+                        }
                         <img src="img/save-the-children.png" alt="Charity logo">
                         <h2>@listing.Title</h2>
                         <p>@listing.Description</p>
